foldEnrichment calculates fold enrichment of a domain (or other feature) among interacting partners of a protein

foldEnrichment calculates fold enrichment of a domain (or other feature) among interacting partners of a protein

foldEnrichment(net, protein_annot, frequency = T)

Arguments

net
data.table that specifies the PPI network: 1. first column specifies a protein ("IDs_interactor_viral") 2. second column specifies it's interacting partner which annotations are of interest ("IDs_interactor_human") 3. third column specifies the degree of a protein in the first column ("IDs_interactor_viral_degree")

protein_annot
data.table that specifies annotations for a protein in the second column of net 1. first column specifies a protein ("IDs_interactor_human") 2. second column specifies features of that protein ("IDs_domain_human") 3. third column specifies background frequency of those features ("domain_frequency)

frequency
fold enrichment or frequency in a set (if TRUE - frequency)

Value

data.table containing fold enrichment or frequency for each domain - protein pair. Columns: IDs_interactor_viral, IDs_interactor_human, IDs_domain_human, domain_frequency_per_set or fold_enrichment