Filter molecular interactions by interaction detection methods and / or participant identification method

filter molecular interaction data by Interaction_detection_methods and / or participant Identification_method

subsetMITABbyMethod(MITABdata, Interaction_detection_methods = NULL,
  Identification_method = NULL,
  Identification_method_participant_A = NULL,
  Identification_method_participant_B = NULL,
  ontology_directory = "./", ontology_date = NULL,
  inverse_filter = F)

Arguments

MITABdata
object of any clean_MItab class (the class that contains "clean_MItab" in it's name and is initially produced by MITABdata)

Interaction_detection_methods
character or character vector, Molecular Interaction ontology code for Interaction detection method, such as "MI:0018", default option results in no filtering by Interaction detection method. Details http://purl.obolibrary.org/obo/MI_0001

Identification_method
character or character vector, Molecular Interaction ontology code for Identification method, such as "MI:0433". Applies the same Identification method filter to both participant_A and participant_B. Default option results in no filtering by Identification method. Details http://purl.obolibrary.org/obo/MI_0002

Identification_method_participant_A
character or character vector, Molecular Interaction ontology code for Identification method for participant A, such as "MI:0433", default option results in no filtering by Interaction detection method for this participant. Parameter is ignored if Identification_method specified. Details http://purl.obolibrary.org/obo/MI_0002

Identification_method_participant_B
character or character vector, Molecular Interaction ontology code for Identification method for participant B, such as "MI:0433", default option results in no filtering by Interaction detection method for this participant. Parameter is ignored if Identification_method specified. Details http://purl.obolibrary.org/obo/MI_0002

ontology_directory
where local copy of the MI ontology is stored (if file not found it will be downloaded)

ontology_date
character, such as "2017_08_25", use specific version of ontology (local copy must be present), by default (NULL) download the latest version. The function doesn't require internet access if this argument is provided.

inverse_filter
logical, inverse filtering criteria

Value

object of the same class as MITABdata, a subset of MITABdata that contains interactions for Interaction_detection_methods and / or participant Identification_method

See also

Examples

{
# Download all human interaction data
full = fullInteractome(taxid = "9606", database = "IntActFTP",
         clean = TRUE, protein_only = TRUE)

# subset two-hybrid interactions
two_hybrid = subsetMITABbyMethod(MITABdata = full,
               Interaction_detection_methods = "MI:0018")

# subset all interactions but two-hybrid
NOT_two_hybrid = subsetMITABbyMethod(MITABdata = full,
                   Interaction_detection_methods = "MI:0018", inverse_filter = T)

# subset affinity purification - mass spectrometry interactions
AP_MS = subsetMITABbyMethod(MITABdata = full,
               Interaction_detection_methods = "MI:0004",  Identification_method = "MI:0433")
}

#> ... looking for the date of the latest IntAct release ...

#> ... looking for the date of the latest IntAct release ...

#> ... loading local copy ...

#> Warning: Found and resolved improper quoting out-of-sample. First healed line 25383: <<uniprotkb:P16054	uniprotkb:Q05769	intact:EBI-298451	intact:EBI-298933|uniprotkb:Q543K3	psi-mi:kpce_mouse(display_long)|uniprotkb:Prkce(gene name)|psi-mi:Prkce(display_short)|uniprotkb:Pkce(gene name synonym)|uniprotkb:Pkcea(gene name synonym)|uniprotkb:nPKC-epsilon(gene name synonym)	psi-mi:pgh2_mouse(display_long)|uniprotkb:Ptgs2(gene name)|psi-mi:Ptgs2(display_short)|uniprotkb:Cox-2(gene name synonym)|uniprotkb:Cox2(gene name synonym)|uniprotkb:Pghs-b(gene name synonym)|uniprotkb:Tis10(gene na>>. If the fields are not quoted (e.g. field separator does not appear within any field), try quote="" to avoid this warning.

#> downloading MI ontology from http://purl.obolibrary.org/obo/mi.obo

#> loading local copy of MI ontology

#> loading local copy of MI ontology